Propagate correctly the dtype of the mask

+++ gammapy/maps/geom.py
@@ -611,4 +611,4 @@ class Geom(abc.ABC):
 
         mask = (energy_edges[:-1] >= energy_min) & (energy_edges[1:] <= energy_max)
         data = np.broadcast_to(mask, shape=self.data_shape)
-        return Map.from_geom(geom=self, data=data)
+        return Map.from_geom(geom=self, data=data, dtype=data.dtype)
